Prevent modification of attribute defaults
Either freeze them, in the case of Strings or empty Hash/Array,
or call a proc to return a unique instance of a mutable object
(mostly relevant for the 'principals' attribute of PermissionTarget).

Receiving a FrozenError in this case may be unexpected for users,
but it will force them into the preferred method of setting attributes
(i.e. using the "#{name}=" setter).

Fixes issue #174.

Signed-off-by: Matthew Blythe <mblythester+git@gmail.com>